% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/agree.coeff2.r
\name{bp2.table}
\alias{bp2.table}
\title{Brennan-Prediger coefficient}
\usage{
bp2.table(ratings, weights = diag(ncol(ratings)), conflev = 0.95, N = Inf,
  print = TRUE)
}
\arguments{
\item{ratings}{a \code{q x q} contingency table showing the distribution of
subjects by rater where \code{q} is the number of categories}

\item{weights}{a vector of weights the same length as the number of
categories}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Brennan-Prediger coefficient and standard error for two raters when
data is a contingency table.
}
\author{
Kilem L. Gwet
}

