% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{disease.l}
\alias{disease.l}
\title{Disease severity rating of 27 patients by four ophthalmologists per person}
\format{An object of class \code{data.frame} containing 108 observations and
3 variables:

\tabular{rll}{
\tab \code{patient} \tab patient number \cr
\tab \code{rater} \tab rater number \cr
\tab \code{category} \tab disease severity \cr
}}
\usage{
disease.l
}
\description{
Sample data in which 27 patients with a particular retinopathy had the
severity of their retinopathy rated by four ophthalmologists.
}
\seealso{
\code{\link{disease.w}}
}
\keyword{data}
\keyword{disease}

