% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ragree-package.R
\docType{data}
\name{disease.w}
\alias{disease.w}
\title{Disease severity rating of 27 patients by four ophthalmologists per person}
\format{An object of class \code{data.frame} containing 27 observations and
4 variables:

\tabular{rll}{
\tab \code{rater1} \tab opthalmologist 1 ratings \cr
\tab \code{rater2} \tab opthalmologist 2 ratings \cr
\tab \code{rater3} \tab opthalmologist 3 ratings \cr
\tab \code{rater4} \tab opthalmologist 4 ratings \cr
}}
\usage{
disease.w
}
\description{
Sample data in which 27 patients with a particular retinopathy had the
severity of their retinopathy rated by four ophthalmologists.
}
\seealso{
\code{\link{disease.l}}
}
\keyword{datasets}

