% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/agree.coeff3.dist.r
\name{fleiss.kappa.dist}
\alias{fleiss.kappa.dist}
\title{Fleiss's generalized kappa coefficient}
\usage{
fleiss.kappa.dist(ratings, weights = "unweighted", conflev = 0.95,
  N = Inf, print = TRUE)
}
\arguments{
\item{ratings}{an \code{n x q} matrix showing the number of raters by
subject and category where \code{n} is the number of subjects and \code{q}
is the number of categories}

\item{weights}{optional weighting}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Fleiss's kappa coefficient and standard error for multiple raters
when data is an \code{n x q} matrix representing the distribution of
raters be subject and by category.
}
\details{
A typical entry associated with a subject and a category, represents the
number of raters who classified the subject into the specified category.
Excludes all subjects that are not rated by any rater.
}
\author{
Kilem L. Gwet
}
\references{
Fleiss, J. L. (1981). Statistical Methods for Rates and
Proportions. John Wiley & Sons.
}

