% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwet_agree.coeff3.raw.r
\name{fleiss.kappa.raw}
\alias{fleiss.kappa.raw}
\title{Fleiss's generalized kappa coefficient}
\usage{
fleiss.kappa.raw(ratings, weights = "unweighted", conflev = 0.95, N = Inf,
  print = TRUE)
}
\arguments{
\item{ratings}{an \code{n x r} matrix of raw alphanumeric ratings from
\code{n} subjects and \code{r} raters}

\item{weights}{optional weighting; see details}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Fleiss's kappa coefficient and standard error for multiple raters
when data is an \code{n x r} matrix of alphanumeric ratings from \code{n}
subjects and \code{r} raters, excluding all subjects that are not rated by
any rater.
}
\details{
\code{weight} is an option matrix of weights or one of \code{"quadratic"},
\code{"linear"}, \code{"ordinal"}, \code{"radical"}, \code{"ratio"},
\code{"circular"}, \code{"bipolar"}, or \code{"unweighted"} for an identity
matrix.
}
\author{
Kilem L. Gwet
}
\references{
Fleiss, J. L. (1981). Statistical Methods for Rates and Proportions. John
Wiley and Sons.
}

