% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_stats.R
\name{gkappa}
\alias{gkappa}
\title{Generalized kappa}
\usage{
gkappa(ratings, data, id = "id", rater = "rater", score = "score")
}
\arguments{
\item{ratings}{\code{n * m} matrix or data frame; \code{n} subjects, 
\code{m} raters}

\item{data}{data in long format with three columns: ids, raters, scores}

\item{id, rater, score}{column names corresponding to IDs, raters, and scores}
}
\value{
A list containing the generalized kappa coefficient, the estimated
variance, the test statistic, and corresponding p-value comparing the 
statistic to a standard normal distribution.

\tabular{rlll}{
\tab \code{$method} \tab agreement method \cr
\tab \code{$ragree.name} \tab method type \cr
\tab \code{$subjects} \tab number of subjects \cr
\tab \code{$raters} \tab number of raters \cr
\tab \code{$categories} \tab number of categories \cr
\tab \code{$value} \tab generalized kappa statistic \cr
\tab \code{$variance} \tab variance \cr
\tab \code{$statistic} \tab test statistic \cr
\tab \code{$stat.name} \tab test statistic name \cr
\tab \code{$p.value} \tab p-value comparing statistic to standard normal 
distribution \cr \cr
\tab \code{$detail} \tab  \cr
\tab \code{...$p_hat_j} \tab proportion of all classifications assigned to 
category j; the pairwise agreement \cr
\tab \code{...$p_bar} \tab overall agreement \cr
\tab \code{...$p_bar_e} \tab expected chance agreement \cr
}
}
\description{
Calculates generalized kappa for k-raters
}
\examples{
## long format using data argument
gkappa(data = disease.l, id = 'patient', rater = 'rater', score = 'category')

## wide format using ratings argument
gkappa(ratings = disease.w)

## example from irr package
library('irr')
data(diagnoses)

gkappa(ratings = diagnoses)

## fleiss's variance calculation is slightly different
kappam.fleiss(diagnoses)

}
\author{
Robert Redd \email{rredd@jimmy.harvard.edu}
}
