% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwet_agree.coeff3.dist.r
\name{gwet.ac1.dist}
\alias{gwet.ac1.dist}
\title{Gwet's AC1/AC2 coefficient}
\usage{
gwet.ac1.dist(ratings, weights = "unweighted", conflev = 0.95, N = Inf,
  print = TRUE)
}
\arguments{
\item{ratings}{an \code{n x q} matrix showing the number of raters by
subject and category where \code{n} is the number of subjects and \code{q}
is the number of categories}

\item{weights}{optional weighting}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Gwet's AC1/AC2 coefficient and standard error for multiple raters
when data is an \code{n x q} matrix representing the distribution of 
raters be subject and by category.
}
\details{
A typical entry associated with a subject and a category, represents the 
number of raters who classified the subject into the specified category. 
Excludes all subjects that are not rated by any rater.
}
\references{
Gwet, K. L. (2008). "Computing inter-rater reliability and its
variance in the presence of high agreement." British Journal of Mathematical
and Statistical Psychology, 61, 29 - 48.
}
\author{
Kilem L. Gwet
}
