% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwet.ac1.raw}
\alias{gwet.ac1.raw}
\title{Gwet's AC1/AC2 coefficient}
\usage{
gwet.ac1.raw(ratings, weights = "unweighted", conflev = 0.95, N = Inf,
  print = TRUE)
}
\arguments{
\item{ratings}{an \code{n x r} matrix of raw alphanumeric ratings from
\code{n} subjects and \code{r} raters}

\item{weights}{optional weighting; see details}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Gwet's AC1/AC2 coefficient and standard error for multiple raters
when data is an \code{n x r} matrix of alphanumeric ratings from \code{n}
subjects and \code{r} raters, excluding all subjects that are not rated by
any rater.
}
\details{
\code{weight} is an option matrix of weights or one of \code{"quadratic"},
\code{"linear"}, \code{"ordinal"}, \code{"radical"}, \code{"ratio"},
\code{"circular"}, \code{"bipolar"}, or \code{"unweighted"} for an identity
matrix.
}
\author{
Kilem L. Gwet
}
\references{
Gwet, K. L. (2008). "Computing inter-rater reliability and its
variance in the presence of high agreement." British Journal of Mathematical
and Statistical Psychology, 61, 29 - 48.
}

