% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwet_agree.coeff3.raw.r
\name{krippen.alpha.raw}
\alias{krippen.alpha.raw}
\title{Krippendorff's alpha coefficient}
\usage{
krippen.alpha.raw(ratings, weights = "unweighted", conflev = 0.95,
  N = Inf, print = TRUE)
}
\arguments{
\item{ratings}{an \code{n x r} matrix of raw alphanumeric ratings from
\code{n} subjects and \code{r} raters}

\item{weights}{optional weighting; see details}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Krippendorff's alpha coefficient and standard error for multiple
raters when data is an \code{n x r} matrix of alphanumeric ratings from
\code{n} subjects and \code{r} raters, excluding all subjects that are not
rated by any rater.
}
\details{
\code{weight} is an option matrix of weights or one of \code{"quadratic"},
\code{"linear"}, \code{"ordinal"}, \code{"radical"}, \code{"ratio"},
\code{"circular"}, \code{"bipolar"}, or \code{"unweighted"} for an identity
matrix.

The algorithm used to compute Krippendorff's alpha is very different from
anything that was published on this topic. Instead, it follows the equations
presented by K. Gwet (2010).
}
\author{
Kilem L. Gwet
}
\references{
Gwet, K. (2012). Handbook of Inter-Rater Reliability: the Definitive Guide
to Measuring the Extent of Agreement among Multiple Raters, 3rd Edition.
Advanced Analytics, LLC; 3rd edition (March 2, 2012).

Krippendorff (1970). "Bivariate agreement coefficients for reliability of
data." Sociological Methodology, 2, 139-150.

Krippendorff (1980). Content analysis: An introduction to its methodology
(2nd ed.), New-bury Park, CA: Sage.
}

