% Generated by roxygen2 (4.0.1): do not edit by hand
\name{krippen2.table}
\alias{krippen2.table}
\title{Krippendorff's alpha coefficient}
\usage{
krippen2.table(ratings, weights = diag(ncol(ratings)), conflev = 0.95,
  N = Inf, print = TRUE)
}
\arguments{
\item{ratings}{a \code{q x q} contingency table showing the distribution of
subjects by rater where \code{q} is the number of categories}

\item{weights}{a vector of weights the same length as the number of
categories}

\item{conflev}{confidence level}

\item{N}{used as denominator in finite population correction}

\item{print}{logical; if \code{TRUE}, prints a summary of the agreement}
}
\description{
Computes Krippendorff's alpha coefficient and standard error for two raters
when data is a contingency table.
}
\author{
Kilem L. Gwet
}

