% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{lr_test}
\alias{lr_test}
\title{LR chi-squared test for two-way tables}
\usage{
lr_test(x, print = TRUE)
}
\arguments{
\item{x}{two-way table}

\item{print}{logical; if \code{TRUE}, prints summary of test}
}
\description{
Likelihood ratio chi-squared test for two-way contingency tables.
}
\examples{
data(Titanic)
dat <- margin.table(Titanic, c(2, 4))
lr_test(dat)

## compare to
loglin(dat, margin = 1:2)

}
\references{
\url{http://ww2.coastal.edu/kingw/statistics/R-tutorials/loglin.html}
}
