% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_stats.R
\name{maxwell}
\alias{maxwell}
\alias{maxwelln}
\title{Maxwell's RE}
\usage{
maxwell(ratings)

maxwelln(ratings)
}
\arguments{
\item{ratings}{an \code{n x 2} matrix or data frame with \code{n} subjects
and 2 raters with binary ratings

note if \code{ratings} contains more than two columns, the first two are
used and the remaining are ignored; to compute the pairwise coefficient for
an \code{n x r} matrix with \code{r} raters, use \code{maxwelln}; see
examples}
}
\description{
Computes Maxwell's RE coefficient for binary data of two raters.
}
\examples{
anx <- +(anxiety > 1)
maxwell(anx)

## compare
irr::maxwell(anx[, -3])

## to get RE for all pairs of raters
maxwelln(anx)

}
\references{
Maxwell, A.E. (1977). Coefficients of agreement between observers and their
interpretation. \emph{Br J Psychiatry} \strong{130}: 79-83.
}
\seealso{
\code{\link[irr]{maxwell}}
}
