% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ragree}
\alias{print.ragree}
\title{ragree class print method}
\usage{
print(x, ...)
}
\arguments{
\item{x}{object of class 'ragree'}

\item{...}{other arguments passed to or from other methods}
}
\description{
ragree class print method
}

