% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ragree-package}
\alias{ragree}
\alias{ragree-package}
\title{Rater agreement}
\description{
Rater agreement
}
\details{
Various measures for quantifying rater agreement
}
\author{
\email{rredd@jimmy.harvard.edu}
}
\keyword{agreement}
\keyword{kappa}
\keyword{rater}
\keyword{unalike}
\keyword{unalikeability}

