% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_stats.R
\name{unalike}
\alias{unalike}
\alias{unalike.default}
\alias{unalike.matrix}
\alias{unalike.data.frame}
\title{Coefficient of unalikeability}
\usage{
unalike(x, ...)

\method{unalike}{default}(x, ...)

\method{unalike}{matrix}(x, ...)

\method{unalike}{data.frame}(x, ..., id, rater, score, summary = TRUE,
  plot = FALSE)
}
\arguments{
\item{x}{a vector of categorical data

alternatively, an \code{n x r} matrix of \code{n} subjects by \code{r}
raters or an \code{n x m} data frame of \code{n} subjects; when \code{x}
is a data  frame, the user must specify which columns correspond to
\code{id}, \code{rater}, and \code{score}}

\item{...}{additional arguments passed to or from other methods}

\item{id}{column name corresponding to IDs}

\item{rater}{column name corresponding to raters}

\item{score}{column name corresponding to rater scores}

\item{summary}{logical; if \code{TRUE}, prints summary statistics for the
unalikeability coefficients}

\item{plot}{logical; if \code{TRUE}, prints a heat map of unalikeability
coefficients; see details}
}
\value{
A list containing the following:

\tabular{rlll}{
\tab \code{$method} \tab agreement method \cr
\tab \code{$ragree.name} \tab method type \cr
\tab \code{$subjects} \tab number of subjects \cr
\tab \code{$raters} \tab number of raters \cr
\tab \code{$categories} \tab number of categories \cr
\tab \code{$value} \tab mean of all unalikeability coefficients \cr
\tab \code{$zzz} \tab a data frame with the summary information printed
when \code{summary = TRUE} \cr
\tab \code{$dat} \tab a long data frame with all coefficients \cr
}
}
\description{
Function to calculate the unalikeability coefficient to quantify the amount 
of variability in categorical data.
}
\details{
The coefficient of unalikeability describes a concept of variability for 
categorical variables and provides a quantitative method for its 
measurement. A smaller coefficient is better corresponding to less variation
in the scores.

For the case of a finite number of observations (\code{n}), a finite number 
of categories (\code{m}) and a finite number of objects, \eqn{k_i}, within
category \code{i}, will allow expression of the coefficient of unalikeablity
as:

\deqn{u = 1 - \sum p_i ^ 2} where \eqn{p_i = k_i / n}.

The interpretation of \eqn{u} is that it represents the proportion of 
possible comparisons (pairings) which are unalike. Note that \eqn{u}
includes comparisons of each response with itself.

If \code{plot} is \code{TRUE}, a \code{\link[ggplot2]{ggplot}} heatmap will
be printed.
}
\examples{
unalike(1, 2)
unalike(rep(1, 10))


## examples in Kader (2007):
l <- list(
  group1 = rep(c('A', 'B'), c(7, 3)),
  group2 = rep(c('A', 'B'), c(5, 5)),
  group3 = rep(c('A', 'B'), c(1, 9)),
  group4 = rep(c('A', 'B', 'C'), c(2, 3, 5))
)

sapply(l, unalike)


## matrix/data frames are assumed to be subjects x raters
mat <- do.call('cbind', l)
unalike(mat) ## see Kader


library('irr')
data(diagnoses)

kappam.fleiss(diagnoses)
unalike(as.matrix(diagnoses))

library('ggplot2')
unalike(as.matrix(diagnoses), plot = TRUE)

}
\references{
Kader, GD. Variability for Categorical Variables. \emph{Journal 
of Statistics Education}, Vol. 15, No. 2: 2007.
}
\author{
Robert Redd \email{rredd@jimmy.harvard.edu}
}
