% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ba_plot}
\alias{ba_plot}
\title{Bland-Altman plot}
\usage{
ba_plot(x, y, xlab = "", ylab = "", xlim = NULL, ylim = NULL,
  panel.first = grid(), ..., line_pars = list(col = c("dodgerblue2",
  "firebrick"), lty = "dashed", lwd = 2, cex = 1.5))
}
\arguments{
\item{x,y}{measurements}

\item{xlab,ylab}{x- and y-axis labels}

\item{xlim,ylim}{x- and y-axis limits}

\item{panel.first}{an expression to be evaluated after the plot axes are
set up but before any plotting takes place (default is just a \code{grid});
see \code{\link{plot.default}}}

\item{...}{additional parameters passed to \code{\link{plot}}}

\item{line_pars}{additional graphical parameters passed to \code{link{text}}
or \code{\link{abline}}; all parameters will be passed as-is except
\code{col} which should be a vector of length one, two, or three to color
the lines and text accordingly; see examples}
}
\description{
Creates a scatter plot using the method of Bland and Altman where the x-
and y-axes are the means of and differences between \code{x} and \code{y},
respectively. Lines for the mean and \code{+/-1.96} times the standard
deviation of the difference between \code{x} and \code{y} are shown.
}
\examples{
set.seed(1)
x <- rnorm(100)
y <- rnorm(100, .5)
ba_plot(x, y, pch = 19)

ba_plot(x, y, col = c('black','green')[(abs(x - y) > 2) + 1],
        panel.first = NULL, pch = '.', cex = 5,
        line_pars = list(col = c('yellow', 'red')))

ba_plot(x, y, cex = 2,
        col = as.numeric(cut(x - y, c(-Inf, -2, 2, Inf))) + 2,
        panel.first = {
          p <- par('usr')
          rect(p[1], p[3], p[2], p[4], col = 'cornsilk')
          grid(col = 'grey90', lty = 1)
        },
        line_pars = list(col = 3:5, family = 'HersheySerif'))
}
\references{
Bland, Altman. Sometime.
}

