% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ba_plot}
\alias{ba_plot}
\title{Bland-Altman plot}
\usage{
ba_plot(x, y, trend = c("loess", "gam", "none"), span = 0.5,
  trend.n = 1000L, alpha = 0.05, xlab = "", ylab = "",
  xlim = NULL, ylim = NULL, panel.first = NULL, panel.last = NULL,
  ..., line_pars = list(col = 1:2, lty = c(2, 1, 2)))
}
\arguments{
\item{x, y}{measurements}

\item{trend}{character string giving the type of trend line, one of
\code{"loess"} (default), \code{"gam"} (requires the \pkg{\link{mgcv}}
package), or \code{"none"} where no curve is drawn}

\item{span}{if \code{trend = "loess"}, a numeric value that controls the
degree of smoothing (see \code{\link{loess}})}

\item{trend.n}{integer value for the number of data points to be used by
\code{\link{predict.loess}}}

\item{alpha}{confidence level affecting the critical values and trend
confidence bands}

\item{xlab, ylab}{x- and y-axis labels}

\item{xlim, ylim}{x- and y-axis limits}

\item{panel.first}{an expression to be evaluated after the plot axes are
set up but before any plotting takes place (default is just a \code{grid});
see \code{\link{plot.default}}}

\item{panel.last}{an expression to be evaluated after plotting has taken
place but before the axes, title, and box are added; see comments about
\code{panel.first}}

\item{...}{additional parameters passed to \code{\link{plot}}}

\item{line_pars}{additional graphical parameters passed to \code{\link{text}}
or \code{\link{abline}}; all parameters will be passed as-is except
\code{col} which should be a vector of length one, two, or three to color
the lines and text accordingly; see examples}
}
\description{
Creates a scatter plot using the method of Bland and Altman where the x-
and y-axes are the means of and differences between \code{x} and \code{y},
respectively. Lines for the mean and \code{+/-1.96} times the standard
deviation of the difference between \code{x} and \code{y} are shown.
}
\examples{
set.seed(1)
x <- rnorm(100)
y <- rnorm(100, 0.5)

## basic usage
ba_plot(x, y)
ba_plot(x, y, trend = 'gam')
ba_plot(x, y, trend = 'none', alpha = 0.1)


## use line_pars to set graphical parameters of individual lines
ba_plot(x, y, col = c('black', 'green')[(abs(x - y) > 2) + 1L],
        pch = '.', cex = 5, line_pars = list(col = c('yellow', 'red')))
        
## panel.first/panel.last is used as in plot.default
ba_plot(x, y, cex = 2, las = 1,
        col = as.numeric(cut(x - y, c(-Inf, -2, 2, Inf))) + 2,
        panel.first = {
          p <- par('usr')
          rect(p[1L], p[3L], p[2L], p[4L], col = 'cornsilk')
          grid(col = 'grey90', lty = 1)
        },
        line_pars = list(col = 3:5, family = 'HersheySerif'))

}
\references{
Altman DG, Bland JM (1983). Measurement in medicine: the analysis of method
comparison studies. \emph{The Statistician} \strong{32}: 307-317.
}
