% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zxx.R
\name{print.ragree}
\alias{print.ragree}
\title{ragree class print method}
\usage{
\method{print}{ragree}(x, ...)
}
\arguments{
\item{x}{object of class 'ragree'}

\item{...}{other arguments passed to or from other methods}
}
\description{
ragree class print method
}

