% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unalike}
\alias{unalike}
\title{Coefficient of unalikeability}
\usage{
unalike(...)

## S3 method for class 'matrix':
unalike(x, ...)

## S3 method for class 'data.frame':
unalike(x, id, rater, score, summary = TRUE, plot = FALSE)
}
\arguments{
\item{...}{numeric values or vector representing the count of each variable}

\item{x}{an \code{n x r} matrix of \code{n} subjects by \code{r} raters or
an \code{n x m} data frame of \code{n} subjects; when \code{x} is a data
frame, the user must specify which columns correspond to \code{id},
\code{rater}, and \code{score}}

\item{id}{data frame column name corresponding to IDs}

\item{rater}{data frame column name corresponding to raters}

\item{score}{data frame column name corresponding to rater scores}

\item{summary}{logical; if \code{TRUE}, prints summary statistics for the
unalikeability coefficients}

\item{plot}{logical; if \code{TRUE}, prints a heat map of unalikeability
coefficients; see details}
}
\value{
A list containing the following:

\tabular{rlll}{
\tab \code{$method} \tab agreement method \cr
\tab \code{$ragree.name} \tab method type \cr
\tab \code{$subjects} \tab number of subjects \cr
\tab \code{$raters} \tab number of raters \cr
\tab \code{$categories} \tab number of categories \cr
\tab \code{$value} \tab mean of all unalikeability coefficients \cr
\tab \code{$zzz} \tab a data frame with the summary information printed
when \code{summary = TRUE} \cr
\tab \code{$dat} \tab a long data frame with all coefficients \cr
}
}
\description{
Function to calculate the unalikeability coefficient to quantify the amount
of variability in categorical data.
}
\details{
The coefficient of unalikeability describes a concept of variability for
categorical variables and provides a quantitative method for its
measurement. A smaller coefficient is better corresponding to less variation
in the scores.

For the case of a finite number of observations (\code{n}), a finite number
of categories (\code{m}) and a finite number of objects, \eqn{k_i}, within
category \code{i}, will allow expression of the coefficient of unalikeablity
as:

\deqn{u = 1 - \sum p_i ^ 2} where \eqn{p_i = k_i / n}.

The interpretation of \eqn{u} is that it represents the proportion of
possible comparisons (pairings) which are unalike. Note that \eqn{u}
includes comparisons of each response with itself.

If \code{plot} is \code{TRUE}, a \code{\link[ggplot2]{ggplot}} heatmap will
be printed.
}
\examples{
## examples in Kader:
## sample data
l <- list(grp1 = c(rep('A', 7), rep('B', 3)),
          grp2 = rep(c('A','B'), each = 5),
          grp3 = c(rep('A', 1), rep('B', 9)))

unalike(1, 9)
sapply(l, function(x) unalike(x))
unalike(sample(c('a','b','c'), 20, replace = TRUE))

mat <- do.call(cbind, l)
colnames(mat) <- c('g1','g2','g3')
unalike(t(mat)) ## see Kader

\dontrun{
library(irr)
data(diagnoses)
kappam.fleiss(diagnoses)
unalike(as.matrix(diagnoses), plot = TRUE)
}
}
\author{
Robert Redd \email{rredd@jimmy.harvard.edu}
}
\references{
Kader, GD. Variability for Categorical Variables. \emph{Journal
of Statistics Education}, Vol. 15, No. 2: 2007.
}

